requirements.txt
inputs
mss
matplotlib
numpy
scikit-image
termcolor
keras
h5py
tensorflow-gpu
Pillow
tensorflow
sklearn
opencv-python
tqdm
vjoy